pyproject.toml
[tool.poetry]
name = "repl_python3_discord-bot-gpt3"
version = "0.1.0"
description = ""
authors = ["Sayuru"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
openai = "^0.10.1"
python-dotenv = "^0.18.0"
Flask = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
